Split the extension API barrel into domain modules.
Extract storage, keys, signing, addresses, and chain reads from the 4k-line
index.js so wallet.js can import leaves instead of cycling through the barrel.
