+++ src/api/extension/index.js
@@ -1,38 +1,29 @@
 import {
-  APIError,
-  DataSignError,
   ERROR,
   EVENT,
   HW,
   LOCAL_STORAGE,
   MAX_TOTAL_ACCOUNTS,
-  NETWORK_ID, NETWORKD_ID_NUMBER,
+  NETWORK_ID,
+  NETWORKD_ID_NUMBER,
   NODE,
   SENDER,
   STORAGE,
   TAB,
   TARGET,
-  TxSignError,
 } from '../../config/config';
-import { POPUP_WINDOW } from '../../config/config';
 import platform from '../../platform';
-import { nativeSafeBinaryBody } from '../../platform/capacitor';
 import { mnemonicToEntropy } from 'bip39';
-import cryptoRandomString from 'crypto-random-string';
 import Loader from '../loader';
 import { createAvatar } from '@dicebear/avatars';
 import { shapes } from '@dicebear/collection';
-// Lazy-load `./wallet` at call sites — a static import cycles (wallet → index)
-// and breaks Jest `requireActual` mocks (e.g. MAX_EXTERNAL_ADDRESS_INDEX).
+// Lazy-load `./wallet` at call sites — a static import used to cycle
+// (wallet → index). wallet.js now imports leaf modules; keep the lazy require
+// so Jest `requireActual` mocks (e.g. MAX_EXTERNAL_ADDRESS_INDEX) stay stable.
 import {
   koiosRequest,
   koiosRequestEnhanced,
-  koiosSubmitTransaction,
-  networkNameToId,
-  utxoFromJson,
   assetsToValue,
-  txToLedger,
-  txToTrezor,
   linkToSrc,
   convertMetadataPropToString,
   extractMetadataImage,
@@ -41,57 +32,124 @@ import {
   Data,
 } from '../util';
 import TransportWebBLE from '@ledgerhq/hw-transport-web-ble';
-import Ada, { HARDENED } from '@cardano-foundation/ledgerjs-hw-app-cardano';
+import Ada from '@cardano-foundation/ledgerjs-hw-app-cardano';
 import AssetFingerprint from '@emurgo/cip14-js';
-import { isAddress } from 'web3-validator';
 import { milkomedaNetworks } from '@dcspark/milkomeda-constants';
-import { Cardano, Serialization } from '@cardano-sdk/core';
-import provider from '../../config/provider';
 import { KOIOS_REQUESTS, addressTxsIndicatesHistory } from '../koios-endpoints';
-import { bigIntLovelace, normalizeLovelaceScalar } from '../lovelace-scalar';
-import { buildVkeyWitnessSet } from '../tx/sign-witness-set';
-import {
-  MAX_COLLATERAL_AMOUNT,
-  isReservedCollateralPresent,
-  parseCollateralAmount,
-  selectCollateralCandidates,
-} from './collateral';
-import {
-  aggregateKoiosUtxosByAddress,
-  aggregateKoiosUtxosToAssets,
-  stakeAddressFromAddressInfo,
-  stakeControlledLovelaceFromAccountInfo,
-  summarizeAddressInfo,
-  summarizeUtxosByAddressEntry,
-} from './stake-balance';
-import {
-  emptyDelegation,
-  normalizeDelegationRow,
-  normalizeStakePool as normalizeStakePoolData,
-} from '../staking';
-import {
-  cacheKey,
-  invalidateAll as invalidateReadCache,
-  withCache,
-} from '../cache';
+import { normalizeLovelaceScalar } from '../lovelace-scalar';
+import { normalizeStakePool as normalizeStakePoolData } from '../staking';
+import { invalidateAll as invalidateReadCache } from '../cache';
 import {
   filterPaymentAddressesForAccountsDisplay,
   getExternalIndices,
   getInternalIndices,
   getUserExternalIndices,
   isMultiAddressEnabled,
-  listEnabledPaymentAddresses,
   normalizeExternalIndices,
   normalizeInternalIndices,
-  deriveExternalPaymentFromAccountPublicKey,
-  derivePaymentFromAccountPublicKey,
-  matchExternalIndicesFromAddresses,
-  matchInternalIndicesFromAddresses,
-  flattenAccountAddressesPayload,
   MAX_EXTERNAL_ADDRESS_INDEX,
   MAX_INTERNAL_ADDRESS_INDEX,
   ADDRESS_ROLE,
 } from './multi-address';
+import {
+  getStorage,
+  setStorage,
+  getCurrentAccountIndex,
+  getNetwork,
+} from './storage';
+import {
+  encryptWithPassword,
+  decryptWithPassword,
+  harden,
+  requestAccountKey,
+} from './keys';
+import {
+  activateDiscoveredExternalAddresses,
+  discoverUsedPaymentIndices,
+  getAddress,
+} from './addresses';
+import {
+  checkCollateral,
+  getBalanceExtended,
+  getTransactions,
+} from './chain-reads';
+
+export {
+  getStorage,
+  setStorage,
+  removeStorage,
+  getCurrentAccountIndex,
+  getNetwork,
+  setNetwork,
+  getCurrentAccount,
+  hasStoredAccounts,
+  getAccounts,
+} from './storage';
+
+export {
+  encryptWithPassword,
+  decryptWithPassword,
+  requestAccountKey,
+  changeWalletPassword,
+} from './keys';
+
+export {
+  isValidEthAddress,
+  extractKeyHash,
+  extractKeyOrScriptHash,
+  verifySigStructure,
+  verifyPayload,
+  verifyTx,
+  signData,
+  signDataCIP30,
+  signTx,
+  signTxHW,
+  submitTx,
+} from './signing';
+
+export {
+  getAddress,
+  getEnabledPaymentAddresses,
+  paymentKeyHashesForSigning,
+  setAccountExternalIndices,
+  setAccountExternalIndicesAt,
+  setAccountInternalIndicesAt,
+  discoverUsedPaymentIndices,
+  discoverUsedExternalIndices,
+  activateDiscoveredExternalAddresses,
+  enableExternalAddressIndex,
+  disableExternalAddressIndex,
+  getRewardAddress,
+  getPubDRepKey,
+  getAccountDRepId,
+  getRegisteredPubStakeKeys,
+  getUnregisteredPubStakeKeys,
+} from './addresses';
+
+export {
+  getDelegation,
+  getPoolMetadata,
+  searchPools,
+  getStakePools,
+  resolveStakeAddressFromPaymentAddress,
+  getAccountStakeAddress,
+  getBalance,
+  getBalanceExtended,
+  getFullBalance,
+  getAccountsControlledStake,
+  getEnabledPaymentAddressDetails,
+  getTransactions,
+  getFiatPrice,
+  getTxInfo,
+  getBlock,
+  getTxUTxOs,
+  getTxMetadata,
+  updateTxInfo,
+  setTxDetail,
+  getSpecificUtxo,
+  getUtxos,
+  getCollateral,
+} from './chain-reads';
 
 export {
   filterPaymentAddressesForAccountsDisplay,
@@ -108,61 +166,9 @@ export {
 
 export const normalizeStakePool = normalizeStakePoolData;
 
-const hasTaggedSets = (cbor) => {
-  const tx = Serialization.Transaction.fromCbor(cbor);
-  return tx.body().hasTaggedSets();
-}
 
-const compareValues = (value1, value2) => {
-  try {
-    const result = value1.checked_sub(value2);
 
-    // If subtraction does not throw and result is not zero, value1 is greater
-    if (!result.is_zero()) {
-      return 1;
-    }
 
-    return 0;
-  } catch (error) {
-    // If we catch an underflow error, value1 is less than value2
-    return -1;
-  }
-}
-
-export const getStorage = (key) => platform.storage.get(key);
-export const setStorage = (item) => platform.storage.set(item);
-export const removeStorage = (item) => platform.storage.remove(item);
-
-export const encryptWithPassword = async (password, rootKeyBytes) => {
-  await Loader.load();
-  const rootKeyHex = rootKeyBytes instanceof Uint8Array
-    ? Buffer.from(rootKeyBytes).toString('hex')
-    : Buffer.from(rootKeyBytes, 'hex').toString('hex');
-  const passwordHex = Buffer.from(password).toString('hex');
-  const salt = cryptoRandomString({ length: 2 * 32 });
-  const nonce = cryptoRandomString({ length: 2 * 12 });
-  return Loader.Cardano.encrypt_with_password(
-    passwordHex,
-    salt,
-    nonce,
-    rootKeyHex
-  );
-};
-
-export const decryptWithPassword = async (password, encryptedKeyHex) => {
-  await Loader.load();
-  const passwordHex = Buffer.from(password).toString('hex');
-  let decryptedHex;
-  try {
-    decryptedHex = Loader.Cardano.decrypt_with_password(
-      passwordHex,
-      encryptedKeyHex
-    );
-  } catch (err) {
-    throw new Error(ERROR.wrongPassword);
-  }
-  return decryptedHex;
-};
 
 // dApp origin allowlist lives in its own module (the trust anchor for the
 // background authZ gate + content-script proxy). Re-exported here to keep the
@@ -205,1586 +211,8 @@ export const syncGlowEffectsDom = (glowEffects) => {
   );
 };
 
-export const getDelegation = async ({ force = false } = {}) => {
-  const network = await getNetwork();
-  const stakeAddress = await getRewardAddress();
-  return withCache(
-    cacheKey('delegation', network?.id, stakeAddress),
-    () => fetchDelegation(stakeAddress),
-    { force }
-  );
-};
 
-const fetchDelegation = async (stakeAddress) => {
-  const request = KOIOS_REQUESTS.getAccountInfo(stakeAddress);
-  const stake = await koiosRequest(request.endpoint, {}, request.body);
 
-  if (!stake || stake.error || !Array.isArray(stake) || !stake[0]) {
-    return emptyDelegation(stakeAddress);
-  }
-
-  const stakeRow = stake[0];
-  const delegation = normalizeDelegationRow(stakeRow, stakeAddress);
-
-  if (!stakeRow.pool_id) {
-    return delegation;
-  }
-
-  const poolRequest = KOIOS_REQUESTS.getPoolInfo([stakeRow.pool_id]);
-  const poolResponse = await koiosRequest(
-    poolRequest.endpoint,
-    {},
-    poolRequest.body
-  );
-
-  if (
-    !poolResponse ||
-    poolResponse.error ||
-    !Array.isArray(poolResponse) ||
-    poolResponse.length === 0
-  ) {
-    return delegation;
-  }
-
-  const pool = normalizeStakePool(poolResponse[0], stakeRow.pool_id);
-  return {
-    ...delegation,
-    poolId: pool.poolId,
-    poolIdHex: pool.poolIdHex,
-    ticker: pool.ticker,
-    description: pool.description,
-    name: pool.name,
-    homepage: pool.homepage,
-    margin: pool.margin,
-    fixedCost: pool.fixedCost,
-    pledge: pool.pledge,
-    activeStake: pool.activeStake,
-    liveSaturation: pool.liveSaturation,
-    blocks: pool.blocks,
-    status: pool.status,
-  };
-};
-
-export const getPoolMetadata = async (poolId) => {
-  if (!poolId) {
-    throw new Error('poolId argument not provided');
-  }
-
-  const request = KOIOS_REQUESTS.getPoolInfo([poolId]);
-  const response = await koiosRequest(request.endpoint, {}, request.body);
-
-  if (!response || response.error || !Array.isArray(response) || response.length === 0) {
-    throw new Error(response?.message || 'Stake pool not found');
-  }
-
-  const poolData = response[0];
-  const pool = normalizeStakePool(poolData, poolId);
-
-  return {
-    ...pool,
-    id: pool.poolId,
-    hex: pool.poolIdHex,
-  };
-};
-
-export const searchPools = async (query) => {
-  if (!query) return [];
-  const searchLower = encodeURIComponent(query.trim().toLowerCase());
-
-  // Use Koios PostgREST filtering for server-side search
-  const listEndpoint = `/pool_list?pool_status=eq.registered&or=(ticker.ilike.*${searchLower}*,pool_id_bech32.ilike.*${searchLower}*)&limit=20`;
-  const poolList = await koiosRequest(listEndpoint);
-
-  if (!poolList || poolList.error || !Array.isArray(poolList) || poolList.length === 0) {
-    return [];
-  }
-
-  // Get detailed info for the matches
-  const poolIds = poolList.map(m => m.pool_id_bech32);
-  const infoRequest = KOIOS_REQUESTS.getPoolInfo(poolIds);
-  const detailedPools = await koiosRequest(infoRequest.endpoint, {}, infoRequest.body);
-
-  if (!detailedPools || detailedPools.error || !Array.isArray(detailedPools)) {
-    return [];
-  }
-
-  return detailedPools.map((pool) => normalizeStakePool(pool));
-};
-
-export const getStakePools = async (limit = 25) => {
-  const cappedLimit = Math.max(1, Math.min(Number(limit) || 25, 100));
-  const poolList = await koiosRequest(
-    `/pool_list?pool_status=eq.registered&limit=${cappedLimit}`
-  );
-
-  if (!poolList || poolList.error || !Array.isArray(poolList) || poolList.length === 0) {
-    return [];
-  }
-
-  const poolIds = poolList.map((pool) => pool.pool_id_bech32).filter(Boolean);
-  if (poolIds.length === 0) return [];
-
-  const infoRequest = KOIOS_REQUESTS.getPoolInfo(poolIds);
-  const detailedPools = await koiosRequest(infoRequest.endpoint, {}, infoRequest.body);
-  if (!detailedPools || detailedPools.error || !Array.isArray(detailedPools)) {
-    return [];
-  }
-
-  return detailedPools.map((pool) => normalizeStakePool(pool));
-};
-
-/**
- * Look up the stake/reward address for a payment address via `/address_info`.
- * Used when the wallet has a payment address but no stored rewardAddr yet.
- */
-export const resolveStakeAddressFromPaymentAddress = async (paymentAddr) => {
-  if (!paymentAddr) return null;
-  try {
-    const request = KOIOS_REQUESTS.getAddressInfo(paymentAddr);
-    const result = await koiosRequest(request.endpoint, {}, request.body);
-    if (result?.error) return null;
-    return stakeAddressFromAddressInfo(result);
-  } catch (error) {
-    console.warn(
-      'resolveStakeAddressFromPaymentAddress failed:',
-      error.message || error
-    );
-    return null;
-  }
-};
-
-/**
- * Stake account for the current wallet: prefer stored rewardAddr, otherwise
- * resolve it from the primary payment address through the chain API.
- */
-export const getAccountStakeAddress = async () => {
-  const stored = await getRewardAddress();
-  if (stored) return stored;
-  const paymentAddr = await getAddress();
-  return resolveStakeAddressFromPaymentAddress(paymentAddr);
-};
-
-export const getBalance = async () => {
-  await Loader.load();
-  const stakeAddress = await getAccountStakeAddress();
-  let utxos = [];
-  if (stakeAddress) {
-    // `_extended: true` is required on Koios — otherwise asset_list is null and
-    // native tokens under the stake key are omitted from the CIP-30 Value.
-    const request = KOIOS_REQUESTS.getAccountUtxos(stakeAddress, true);
-    const result = await koiosRequest(request.endpoint, {}, request.body);
-    if (result?.error) {
-      if (result.status_code === 400) throw APIError.InvalidRequest;
-      else if (result.status_code === 500) throw APIError.InternalError;
-    } else if (Array.isArray(result)) {
-      utxos = result;
-    }
-  }
-  // Fallback: enabled payment addresses only (legacy / no stake addr).
-  if (utxos.length === 0 && !stakeAddress) {
-    const paymentAddresses = await getEnabledPaymentAddresses();
-    const addressList = paymentAddresses.map((a) => a.paymentAddr).filter(Boolean);
-    if (addressList.length === 0) {
-      return Loader.Cardano.Value.new(Loader.Cardano.BigNum.from_str('0'));
-    }
-    const request = KOIOS_REQUESTS.getAddressesUtxos(addressList, true);
-    const result = await koiosRequest(request.endpoint, {}, request.body);
-    if (result?.error) {
-      if (result.status_code === 400) throw APIError.InvalidRequest;
-      else if (result.status_code === 500) throw APIError.InternalError;
-      else return Loader.Cardano.Value.new(Loader.Cardano.BigNum.from_str('0'));
-    }
-    utxos = Array.isArray(result) ? result : [];
-  }
-
-  if (utxos.length === 0) {
-    return Loader.Cardano.Value.new(Loader.Cardano.BigNum.from_str('0'));
-  }
-
-  const assets = aggregateKoiosUtxosToAssets(utxos);
-  return await assetsToValue(assets);
-};
-
-export const getBalanceExtended = async ({ force = false } = {}) => {
-  const network = await getNetwork();
-  const stakeAddress = await getAccountStakeAddress();
-  if (!stakeAddress) {
-    const addresses = await getEnabledPaymentAddresses();
-    const addressList = addresses.map((a) => a.paymentAddr).filter(Boolean);
-    if (addressList.length === 0) return [];
-    return withCache(
-      cacheKey(
-        'balance-extended',
-        network?.id,
-        addressList.join(','),
-        addresses.map((a) => `${a.role ?? 0}:${a.index}`).join('-')
-      ),
-      () => fetchBalanceExtended(addressList),
-      { force }
-    );
-  }
-  // Stake-controlled UTxOs include every payment address under the account
-  // (external + change), so the wallet total matches chain controlled_amount.
-  return withCache(
-    cacheKey('balance-extended-stake', network?.id, stakeAddress),
-    () => fetchBalanceFromStake(stakeAddress),
-    { force }
-  );
-};
-
-const fetchBalanceExtended = async (addresses) => {
-  const list = Array.isArray(addresses) ? addresses : [addresses];
-  const request = KOIOS_REQUESTS.getAddressesUtxos(list, true);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-
-  if (result.error) {
-    if (result.status_code === 400) throw APIError.InvalidRequest;
-    else if (result.status_code === 500) throw APIError.InternalError;
-    else return [];
-  }
-
-  if (!result || result.length === 0) {
-    return [];
-  }
-
-  return aggregateKoiosUtxosToAssets(result);
-};
-
-const fetchBalanceFromStake = async (stakeAddress) => {
-  const request = KOIOS_REQUESTS.getAccountUtxos(stakeAddress, true);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-
-  if (result?.error) {
-    if (result.status_code === 400) throw APIError.InvalidRequest;
-    else if (result.status_code === 500) throw APIError.InternalError;
-    else return [];
-  }
-
-  if (!result || result.length === 0) {
-    return [];
-  }
-
-  return aggregateKoiosUtxosToAssets(result);
-};
-
-export const getFullBalance = async () => {
-  const stakeAddress = await getAccountStakeAddress();
-  if (!stakeAddress) return '0';
-
-  const request = KOIOS_REQUESTS.getAccountInfo(stakeAddress);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-
-  if (result?.error || !result?.[0]) return '0';
-  return stakeControlledLovelaceFromAccountInfo(result[0]);
-};
-
-/**
- * Stake-controlled ADA for every stored account (batch `/account_info`).
- * Used by the Accounts list so rows show controlled stake — not primary
- * payment-address contents.
- *
- * @returns {Promise<Record<string, { lovelace: string, status: string|null, poolId: string|null }>>}
- */
-export const getAccountsControlledStake = async () => {
-  const accounts = await getStorage(STORAGE.accounts);
-  const network = await getNetwork();
-  if (!accounts || typeof accounts !== 'object' || !network?.id) {
-    return {};
-  }
-
-  const accountKeysByStake = new Map();
-  for (const key of Object.keys(accounts)) {
-    const rewardAddr = accounts[key]?.[network.id]?.rewardAddr;
-    if (!rewardAddr) continue;
-    const list = accountKeysByStake.get(rewardAddr) || [];
-    list.push(key);
-    accountKeysByStake.set(rewardAddr, list);
-  }
-
-  const stakeAddresses = Array.from(accountKeysByStake.keys());
-  if (stakeAddresses.length === 0) return {};
-
-  const request = KOIOS_REQUESTS.getAccountsInfo(stakeAddresses);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  const out = {};
-
-  if (Array.isArray(result)) {
-    for (const row of result) {
-      const stake = row?.stake_address;
-      if (!stake) continue;
-      const keys = accountKeysByStake.get(stake) || [];
-      const lovelace = stakeControlledLovelaceFromAccountInfo(row);
-      const status =
-        row.status ||
-        (row.registered === true
-          ? 'registered'
-          : row.registered === false
-            ? 'unregistered'
-            : null);
-      const poolId = row.delegated_pool || row.pool_id || null;
-      for (const key of keys) {
-        out[key] = { lovelace, status, poolId };
-      }
-    }
-  }
-
-  for (const keys of accountKeysByStake.values()) {
-    for (const key of keys) {
-      if (out[key] == null) {
-        out[key] = { lovelace: '0', status: null, poolId: null };
-      }
-    }
-  }
-
-  return out;
-};
-
-/**
- * Enabled payment/change addresses for the current account, enriched with
- * per-address contents (ADA, UTxO count, native asset count).
- *
- * @param {{ accountsDisplay?: boolean }} [options] - When `accountsDisplay`,
- *   refresh discovery, prefer stake `/account_utxos` for funded addresses
- *   (so every address holding assets is listed even if prior discovery or
- *   `/address_info` missed it), then filter to assets + user-activated.
- */
-export const getEnabledPaymentAddressDetails = async (options = {}) => {
-  const accountsDisplay = Boolean(options?.accountsDisplay);
-  await Loader.load();
-  const network = await getNetwork();
-  const networkId = NETWORKD_ID_NUMBER[network.name || network.id];
-
-  if (accountsDisplay) {
-    try {
-      const currentIndex = await getCurrentAccountIndex();
-      await activateDiscoveredExternalAddresses(currentIndex, {
-        networkKeys: [network.id],
-      });
-    } catch (error) {
-      console.warn(
-        'Accounts address discovery failed:',
-        error?.message || error
-      );
-    }
-  }
-
-  let currentAccount = await getCurrentAccount();
-  let rows = listEnabledPaymentAddresses(
-    Loader.Cardano,
-    currentAccount,
-    networkId
-  );
-
-  /** @type {Map<string, { lovelace: bigint, utxoCount: number, assetUnits: Set<string> }>} */
-  let fundedByAddr = new Map();
-  if (accountsDisplay && currentAccount.rewardAddr) {
-    try {
-      const utxoReq = KOIOS_REQUESTS.getAccountUtxos(
-        currentAccount.rewardAddr,
-        true
-      );
-      const utxos = await koiosRequest(utxoReq.endpoint, {}, utxoReq.body);
-      if (Array.isArray(utxos)) {
-        fundedByAddr = aggregateKoiosUtxosByAddress(utxos);
-      }
-    } catch (error) {
-      console.warn(
-        'Accounts funded-address scan failed:',
-        error?.message || error
-      );
-    }
-  }
-
-  // Activate any CIP-1852 indices that currently hold UTxOs but were not yet
-  // in externalIndices/internalIndices (common for accounts never soft-refreshed).
-  if (accountsDisplay && currentAccount.publicKey && fundedByAddr.size > 0) {
-    const fundedAddrs = Array.from(fundedByAddr.keys());
-    const extFromFunded = matchExternalIndicesFromAddresses(
-      Loader.Cardano,
-      currentAccount.publicKey,
-      networkId,
-      fundedAddrs
-    );
-    const intFromFunded = matchInternalIndicesFromAddresses(
-      Loader.Cardano,
-      currentAccount.publicKey,
-      networkId,
-      fundedAddrs
-    );
-    const prevExt = getExternalIndices(currentAccount);
-    const prevInt = getInternalIndices(currentAccount);
-    const mergedExt = normalizeExternalIndices([...prevExt, ...extFromFunded]);
-    const mergedInt = normalizeInternalIndices([...prevInt, ...intFromFunded]);
-    const extChanged =
-      mergedExt.length !== prevExt.length ||
-      mergedExt.some((n, i) => n !== prevExt[i]);
-    const intChanged =
-      mergedInt.length !== prevInt.length ||
-      mergedInt.some((n, i) => n !== prevInt[i]);
-    if (extChanged || intChanged) {
-      const currentIndex = await getCurrentAccountIndex();
-      const accounts = await getStorage(STORAGE.accounts);
-      if (accounts?.[currentIndex]) {
-        if (!Array.isArray(accounts[currentIndex].userExternalIndices)) {
-          accounts[currentIndex].userExternalIndices = prevExt;
-        }
-        accounts[currentIndex].externalIndices = mergedExt;
-        accounts[currentIndex].internalIndices = mergedInt;
-        await setStorage({ [STORAGE.accounts]: { ...accounts } });
-        invalidateReadCache();
-        currentAccount = await getCurrentAccount();
-        rows = listEnabledPaymentAddresses(
-          Loader.Cardano,
-          currentAccount,
-          networkId
-        );
-      }
-    }
-  }
-
-  if (rows.length === 0) return [];
-
-  // Prefer stake-UTxO totals for funded addresses; `/address_info` only for the rest
-  // (user-activated empties). Avoids false "empty" rows that the display filter drops.
-  const needInfo = rows
-    .map((r) => r.paymentAddr)
-    .filter((addr) => addr && !fundedByAddr.has(addr));
-  const byAddrInfo = new Map();
-  if (needInfo.length > 0) {
-    try {
-      const infoReq = KOIOS_REQUESTS.getAddressesInfo(needInfo);
-      const infoRows = await koiosRequest(infoReq.endpoint, {}, infoReq.body);
-      if (Array.isArray(infoRows)) {
-        for (const row of infoRows) {
-          if (row?.address) byAddrInfo.set(row.address, row);
-        }
-      }
-    } catch (error) {
-      console.warn(
-        'Accounts address_info enrich failed:',
-        error?.message || error
-      );
-    }
-  }
-
-  const details = rows.map((row) => {
-    const funded = fundedByAddr.get(row.paymentAddr);
-    if (funded) {
-      const summary = summarizeUtxosByAddressEntry(funded);
-      return {
-        ...row,
-        lovelace: summary.lovelace,
-        utxoCount: summary.utxoCount,
-        nativeAssetCount: summary.nativeAssetCount,
-      };
-    }
-    const summary = summarizeAddressInfo(byAddrInfo.get(row.paymentAddr));
-    return {
-      ...row,
-      lovelace: summary.lovelace,
-      utxoCount: summary.utxoCount,
-      nativeAssetCount: summary.nativeAssetCount,
-    };
-  });
-
-  // Stake UTxOs on addresses we could not map to a known index (e.g. missing
-  // account publicKey, or index beyond the scan cap) still belong in the list.
-  if (accountsDisplay && fundedByAddr.size > 0) {
-    const listed = new Set(details.map((row) => row.paymentAddr));
-    for (const [addr, funded] of fundedByAddr) {
-      if (listed.has(addr)) continue;
-      const summary = summarizeUtxosByAddressEntry(funded);
-      details.push({
-        role: ADDRESS_ROLE.external,
-        index: null,
-        paymentAddr: addr,
-        paymentKeyHash: null,
-        lovelace: summary.lovelace,
-        utxoCount: summary.utxoCount,
-        nativeAssetCount: summary.nativeAssetCount,
-      });
-    }
-  }
-
-  if (!accountsDisplay) return details;
-  return filterPaymentAddressesForAccountsDisplay(details, currentAccount);
-};
-
-export const getTransactions = async (paginate = 1, count = 10, { force = false } = {}) => {
-  const network = await getNetwork();
-  const stakeAddress = await getRewardAddress();
-  // The bounded fetch returns the same leading 100 txs regardless of the UI's
-  // local paging cursor, so a single per-account/network cache entry is safe.
-  return withCache(
-    cacheKey('account-txs', network?.id, stakeAddress),
-    () => fetchTransactions(stakeAddress),
-    { force }
-  );
-};
-
-const fetchTransactions = async (stakeAddress) => {
-  const request = KOIOS_REQUESTS.getAccountTxs(stakeAddress, 0);
-  // Bound the Koios direct-path response (mainnet accounts can have thousands of
-  // txs; an unbounded fetch is slow/huge and makes the history "load forever").
-  // `order`/`limit` are PostgREST reserved params Koios honours; the Blockfrost
-  // adapter ignores them and self-caps at 100.
-  const boundedEndpoint = `${request.endpoint}&order=block_height.desc&limit=100`;
-
-  // Never let a hung/slow provider stall the history spinner indefinitely.
-  const controller = new AbortController();
-  const timeout = setTimeout(() => controller.abort(), 20000);
-  let result;
-  try {
-    result = await koiosRequest(
-      boundedEndpoint,
-      {},
-      request.body,
-      controller.signal
-    );
-  } catch (error) {
-    console.warn('getTransactions failed:', error?.message || error);
-    return [];
-  } finally {
-    clearTimeout(timeout);
-  }
-
-  if (!result || result.error) return [];
-  
-  let processedTransactions = result.map(tx => ({
-    txHash: tx.tx_hash,
-    blockHeight: tx.block_height,
-    epochNo: tx.epoch_no,
-    epochSlot: tx.epoch_slot,
-    absoluteSlot: tx.absolute_slot,
-    txTimestamp: tx.tx_timestamp,
-    txBlockIndex: tx.tx_block_index,
-    txSize: tx.tx_size,
-    totalOutput: tx.total_output,
-    fee: tx.fee,
-    deposit: tx.deposit,
-    invalidBefore: tx.invalid_before,
-    invalidAfter: tx.invalid_after,
-    collateralInputs: tx.collateral_inputs,
-    collateralOutput: tx.collateral_output,
-    referenceInputs: tx.reference_inputs,
-    inputs: tx.inputs || [],
-    outputs: tx.outputs || [],
-    withdrawals: tx.withdrawals || [],
-    assetsMinted: tx.assets_minted || [],
-    metadata: tx.metadata,
-    certificates: tx.certificates || [],
-    nativeScripts: tx.native_scripts || [],
-    plutusContracts: tx.plutus_contracts || [],
-    votingProcedures: tx.voting_procedures || [],
-    proposalProcedures: tx.proposal_procedures || []
-  }));
-  
-  return processedTransactions;
-};
-
-/**
- * Cached fiat price lookup. Previously each wallet mount refetched the rate
- * (it lived only in a component ref), so returning to the wallet always
- * re-hit the provider. Cached per-currency with the shared TTL.
- */
-export const getFiatPrice = async (currency, { force = false } = {}) =>
-  withCache(
-    cacheKey('fiat-price', currency),
-    () => provider.api.price(currency),
-    { force }
-  );
-
-export const getTxInfo = async (txHash) => {
-  const request = KOIOS_REQUESTS.getTxInfo(txHash);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  if (!result || result.error || result.length === 0) return null;
-  return result[0];
-};
-
-export const getBlock = async (blockHashOrNumb) => {
-  let request;
-  let result;
-  
-  // Check if it's a block height (number) or block hash (string)
-  if (typeof blockHashOrNumb === 'number' || !isNaN(blockHashOrNumb)) {
-    request = KOIOS_REQUESTS.getBlockByHeight(blockHashOrNumb);
-    result = await koiosRequest(request.endpoint, {}, request.body);
-  } else {
-    request = KOIOS_REQUESTS.getBlockByHash(blockHashOrNumb);
-    result = await koiosRequest(request.endpoint, {}, request.body);
-  }
-  
-  if (!result || result.error || result.length === 0) return null;
-  return result[0];
-};
-
-// Helper function to convert Koios UTXO format to expected format
-const convertKoiosUtxosToExpectedFormat = (koiosUtxos) => {
-  if (!koiosUtxos) return null;
-  const normalizeAddress = (utxo) =>
-    utxo.payment_addr?.bech32 ||
-    utxo.address ||
-    utxo.payment_addr ||
-    utxo.stake_address ||
-    utxo.stake_addr?.bech32 ||
-    utxo.stake_addr ||
-    null;
-  
-  return {
-    inputs: (koiosUtxos.inputs || []).map(input => ({
-      address: normalizeAddress(input),
-      stake_address: input.stake_addr || input.stake_address || input.stake_addr?.bech32,
-      tx_hash: input.tx_hash,
-      tx_index: input.tx_index,
-      value: input.value,
-      asset_list: input.asset_list || [],
-      datum_hash: input.datum_hash,
-      inline_datum: input.inline_datum,
-      reference_script: input.reference_script
-    })),
-    outputs: (koiosUtxos.outputs || []).map(output => ({
-      address: normalizeAddress(output),
-      stake_address: output.stake_addr || output.stake_address || output.stake_addr?.bech32,
-      tx_hash: output.tx_hash,
-      tx_index: output.tx_index,
-      value: output.value,
-      asset_list: output.asset_list || [],
-      datum_hash: output.datum_hash,
-      inline_datum: output.inline_datum,
-      reference_script: output.reference_script
-    }))
-  };
-};
-
-export const getTxUTxOs = async (txHash) => {
-  const request = KOIOS_REQUESTS.getTxUtxos(txHash);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  if (!result || result.error || result.length === 0) return null;
-  
-  // Convert Koios format to expected format
-  const converted = convertKoiosUtxosToExpectedFormat(result[0]);
-  return converted;
-};
-
-export const getTxMetadata = async (txHash) => {
-  const request = KOIOS_REQUESTS.getTxMetadata(txHash);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  if (!result || result.error || result.length === 0) return null;
-  return result[0];
-};
-
-// Helper function to convert Koios transaction format to expected format
-const convertKoiosTxToExpectedFormat = (koiosTx) => {
-  if (!koiosTx) return null;
-  
-  // Calculate transaction type indicators from certificates and other data
-  const certificates = koiosTx.certificates || [];
-  const withdrawals = koiosTx.withdrawals || [];
-  const assetsMinted = koiosTx.assets_minted || [];
-  const plutusContracts = koiosTx.plutus_contracts || [];
-  
-  // Count different types of certificates
-  const delegationCount = certificates.filter(cert => 
-    cert.cert_type === 'delegation' || cert.cert_type === 'deleg_reg'
-  ).length;
-  
-  const stakeCertCount = certificates.filter(cert => 
-    cert.cert_type === 'stake_registration' || cert.cert_type === 'stake_deregistration'
-  ).length;
-  
-  const poolRetireCount = certificates.filter(cert => 
-    cert.cert_type === 'pool_retirement'
-  ).length;
-  
-  const poolUpdateCount = certificates.filter(cert => 
-    cert.cert_type === 'pool_registration' || cert.cert_type === 'pool_update'
-  ).length;
-  
-  // Count other transaction types
-  const withdrawalCount = withdrawals.length;
-  const assetMintOrBurnCount = assetsMinted.length;
-  const redeemerCount = plutusContracts.reduce((count, contract) => 
-    count + (contract.redeemers ? contract.redeemers.length : 0), 0
-  );
-  
-  return {
-    // Basic transaction info
-    tx_hash: koiosTx.tx_hash,
-    block_height: koiosTx.block_height,
-    block_hash: koiosTx.block_hash,
-    epoch_no: koiosTx.epoch_no,
-    epoch_slot: koiosTx.epoch_slot,
-    absolute_slot: koiosTx.absolute_slot,
-    tx_timestamp: koiosTx.tx_timestamp,
-    tx_block_index: koiosTx.tx_block_index,
-    tx_size: koiosTx.tx_size,
-    
-    // Financial info
-    total_output: koiosTx.total_output,
-    fee: koiosTx.fee,
-    treasury_donation: koiosTx.treasury_donation,
-    deposit: koiosTx.deposit,
-    
-    // Validity
-    invalid_before: koiosTx.invalid_before,
-    invalid_after: koiosTx.invalid_after,
-    
-    // UTXOs
-    inputs: koiosTx.inputs || [],
-    outputs: koiosTx.outputs || [],
-    
-    // Additional data
-    collateral_inputs: koiosTx.collateral_inputs,
-    collateral_output: koiosTx.collateral_output,
-    reference_inputs: koiosTx.reference_inputs,
-    withdrawals: koiosTx.withdrawals,
-    assets_minted: koiosTx.assets_minted,
-    certificates: koiosTx.certificates,
-    native_scripts: koiosTx.native_scripts,
-    plutus_contracts: koiosTx.plutus_contracts,
-    
-    // Legacy field names for compatibility
-    fees: koiosTx.fee,
-    valid_contract: true, // Default to true for now
-    
-    // Transaction type detection fields
-    redeemer_count: redeemerCount,
-    withdrawal_count: withdrawalCount,
-    delegation_count: delegationCount,
-    asset_mint_or_burn_count: assetMintOrBurnCount,
-    stake_cert_count: stakeCertCount,
-    pool_retire_count: poolRetireCount,
-    pool_update_count: poolUpdateCount
-  };
-};
-
-export const updateTxInfo = async (txHash) => {
-  const currentAccount = await getCurrentAccount();
-  const network = await getNetwork();
-
-  let detail = await currentAccount[network.id].history.details[txHash];
-
-  if (typeof detail !== 'object' || !detail.info || !detail.block || !detail.utxos || !detail.metadata) {
-    detail = {};
-    
-    // Get transaction info
-    const info = await getTxInfo(txHash);
-    
-    if (info) {
-      // Convert Koios format to expected format
-      detail.info = convertKoiosTxToExpectedFormat(info);
-      
-      // Get block info if we have block height
-      if (info.block_height) {
-        detail.block = await getBlock(info.block_height);
-      }
-    }
-    
-    // Get transaction UTXOs
-    const uTxOs = await getTxUTxOs(txHash);
-    
-    if (uTxOs) {
-      detail.utxos = uTxOs;
-    }
-    
-    // Get transaction metadata
-    const metadata = await getTxMetadata(txHash);
-    if (metadata) {
-      detail.metadata = metadata;
-    }
-  }
-
-  return detail;
-};
-
-export const setTxDetail = async (txObject) => {
-  const currentIndex = await getCurrentAccountIndex();
-  const network = await getNetwork();
-  const accounts = await getStorage(STORAGE.accounts);
-  for (const txHash of Object.keys(txObject)) {
-    const txDetail = txObject[txHash];
-    accounts[currentIndex][network.id].history.details[txHash] = txDetail;
-    await setStorage({
-      [STORAGE.accounts]: {
-        ...accounts,
-      },
-    });
-    delete txObject[txHash];
-  }
-  return true;
-};
-
-export const getSpecificUtxo = async (txHash, txId) => {
-  const request = KOIOS_REQUESTS.getTxUtxos(txHash);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  if (!result || result.error || result.length === 0) return null;
-  return result[0].outputs[txId];
-};
-
-/**
- *
- * @param {string} amount - cbor value
- * @param {Object} paginate
- * @param {number} paginate.page
- * @param {number} paginate.limit
- * @returns
- */
-export const getUtxos = async (amount = undefined, paginate = undefined) => {
-  const currentAccount = await getCurrentAccount();
-  const paymentAddresses = await getEnabledPaymentAddresses();
-  const addressList = paymentAddresses.map((a) => a.paymentAddr).filter(Boolean);
-  const stakeAddress = await getAccountStakeAddress();
-
-  let result;
-  if (stakeAddress) {
-    // Extended UTxOs include asset_list (Koios drops tokens when false).
-    const request = KOIOS_REQUESTS.getAccountUtxos(stakeAddress, true);
-    result = await koiosRequest(request.endpoint, {}, request.body);
-  } else {
-    if (addressList.length === 0) return [];
-    const request = KOIOS_REQUESTS.getAddressesUtxos(addressList, true);
-    result = await koiosRequest(request.endpoint, {}, request.body);
-  }
-
-  if (result?.error) {
-    if (result.status_code === 400) throw APIError.InvalidRequest;
-    else if (result.status_code === 500) throw APIError.InternalError;
-    else return [];
-  }
-
-  let utxos = Array.isArray(result) ? result : [];
-
-  if (currentAccount.collateral) {
-    utxos = utxos.filter(
-      (utxo) =>
-        !(
-          utxo.tx_hash === currentAccount.collateral.txHash &&
-          (utxo.output_index ?? utxo.tx_index) === currentAccount.collateral.txId
-        )
-    );
-  }
-
-  const fallbackOwner = addressList[0] || currentAccount.paymentAddr;
-
-  // Spend only from addresses we can witness (enabled external + change).
-  // Balance aggregation still uses the full stake set via getBalance.
-  const enabledOwners = new Set(addressList.filter(Boolean));
-  if (enabledOwners.size > 0) {
-    utxos = utxos.filter((utxo) =>
-      enabledOwners.has(utxo.address || fallbackOwner)
-    );
-  }
-
-  let convertedUtxos = await Promise.all(
-    utxos.map(async (utxo) => {
-      const owner = utxo.address || fallbackOwner;
-      const formattedUtxo = {
-        tx_hash: utxo.tx_hash,
-        output_index: utxo.output_index ?? utxo.tx_index,
-        amount: [
-          { unit: 'lovelace', quantity: utxo.value || '0' },
-          ...(utxo.asset_list || []).map((asset) => ({
-            unit: asset.policy_id + asset.asset_name,
-            quantity: asset.quantity || '0',
-          })),
-        ],
-      };
-
-      return await utxoFromJson(formattedUtxo, owner);
-    })
-  );
-
-  // filter utxos
-  if (amount) {
-    await Loader.load();
-    let filterValue;
-    try {
-      filterValue = Loader.Cardano.Value.from_bytes(Buffer.from(amount, 'hex'));
-    } catch (e) {
-      throw APIError.InvalidRequest;
-    }
-
-    convertedUtxos = convertedUtxos.filter(
-      (unspent) =>
-        !compareValues(unspent.output().amount(), filterValue) ||
-        compareValues(unspent.output().amount(), filterValue) !== -1
-    );
-  }
-
-  if ((amount || paginate) && convertedUtxos.length <= 0) {
-    return null;
-  }
-  return convertedUtxos;
-};
-
-/**
- * Clear stale reserved collateral when the UTxO is no longer on-chain.
- * Mutates `currentAccount[network.id].collateral` in place.
- * @returns {boolean} true when collateral was cleared (caller should persist)
- */
-const checkCollateral = async (currentAccount, network, checkTx) => {
-  const reserved = currentAccount[network.id].collateral;
-  if (!reserved) return false;
-
-  if (checkTx) {
-    const transactions = await getTransactions();
-    if (
-      transactions.length <= 0 ||
-      currentAccount[network.id].history.confirmed.includes(
-        transactions[0].txHash
-      )
-    ) {
-      return false;
-    }
-  }
-
-  const address = await getAddress();
-  const request = KOIOS_REQUESTS.getAddressInfo(address);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-
-  if (result.error || !result[0]) {
-    if (result.status_code === 400) throw APIError.InvalidRequest;
-    else if (result.status_code === 500) throw APIError.InternalError;
-    else return false;
-  }
-
-  const utxos = result[0].utxo_set || [];
-  if (!isReservedCollateralPresent(utxos, reserved)) {
-    delete currentAccount[network.id].collateral;
-    return true;
-  }
-  return false;
-};
-
-const decodeCollateralCoinCbor = (hex) => {
-  const bytes = Buffer.from(hex, 'hex');
-  try {
-    return BigInt(Loader.Cardano.BigNum.from_bytes(bytes).to_str());
-  } catch (_) {
-    // fall through — some dApps send a CBOR Value instead of a bare Coin
-  }
-  try {
-    return BigInt(Loader.Cardano.Value.from_bytes(bytes).coin().to_str());
-  } catch (_) {
-    throw new Error('could not parse collateral amount');
-  }
-};
-
-/**
- * CIP-30 getCollateral (deprecated; prefer CIP-40 collateral return).
- * @param {{ amount?: string|number }|string|number|undefined} params
- * @returns {Promise<import('@emurgo/cardano-serialization-lib-browser').TransactionUnspentOutput[]|null>}
- */
-export const getCollateral = async (params) => {
-  await Loader.load();
-  const currentIndex = await getCurrentAccountIndex();
-  const accounts = await getStorage(STORAGE.accounts);
-  const currentAccount = accounts[currentIndex];
-  const network = await getNetwork();
-  if (await checkCollateral(currentAccount, network, true)) {
-    await setStorage({ [STORAGE.accounts]: accounts });
-  }
-
-  const amountRaw =
-    params && typeof params === 'object' && !Array.isArray(params)
-      ? params.amount
-      : params;
-
-  let minLovelace;
-  try {
-    minLovelace = parseCollateralAmount(amountRaw, {
-      decodeCoin: decodeCollateralCoinCbor,
-    });
-  } catch (e) {
-    throw {
-      ...APIError.InvalidRequest,
-      info: e?.message || APIError.InvalidRequest.info,
-    };
-  }
-
-  const collateral = currentAccount[network.id].collateral;
-  if (collateral) {
-    const reservedCoin = BigInt(collateral.lovelace.toString());
-    if (reservedCoin >= minLovelace) {
-      return [
-        Loader.Cardano.TransactionUnspentOutput.new(
-          Loader.Cardano.TransactionInput.new(
-            Loader.Cardano.TransactionHash.from_bytes(
-              Buffer.from(collateral.txHash, 'hex')
-            ),
-            parseInt(collateral.txId, 10)
-          ),
-          Loader.Cardano.TransactionOutput.new(
-            Loader.Cardano.Address.from_bech32(
-              currentAccount[network.id].paymentAddr
-            ),
-            Loader.Cardano.Value.new(
-              Loader.Cardano.BigNum.from_str(collateral.lovelace.toString())
-            )
-          )
-        ),
-      ];
-    }
-  }
-
-  const utxos = await getUtxos();
-  if (!utxos || utxos.length <= 0) return null;
-
-  const candidates = utxos.map((utxo) => {
-    const amt = utxo.output().amount();
-    const ma = amt.multiasset();
-    return {
-      coin: BigInt(amt.coin().to_str()),
-      multiassetLen: ma ? ma.len() : 0,
-      utxo,
-    };
-  });
-
-  const selected = selectCollateralCandidates(candidates, minLovelace);
-  if (!selected) {
-    if (minLovelace === MAX_COLLATERAL_AMOUNT && amountRaw == null) {
-      // Back-compat: no amount requested and nothing suitable → empty list
-      // (legacy Nami behavior) rather than null.
-      return [];
-    }
-    return null;
-  }
-  return selected;
-};
-
-export const getAddress = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  // Primary receive address remains external index 0 (CIP-30 / QR default).
-  return currentAccount.paymentAddr;
-};
-
-/**
- * Enabled CIP-1852 external payment addresses for the current account
- * (index 0 plus any Advanced multi-address indices).
- */
-export const getEnabledPaymentAddresses = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  const network = await getNetwork();
-  const networkId = NETWORKD_ID_NUMBER[network.name || network.id];
-  return listEnabledPaymentAddresses(
-    Loader.Cardano,
-    currentAccount,
-    networkId
-  );
-};
-
-/**
- * Payment key hashes for every enabled external + internal address on an
- * account. Used so fee sizing and witnesses cover change/multi-address UTxOs.
- *
- * @param {object} [accountOverride] - network-specific or storage account
- * @returns {Promise<string[]>}
- */
-export const paymentKeyHashesForSigning = async (accountOverride) => {
-  await Loader.load();
-  const account = accountOverride || (await getCurrentAccount());
-  const network = await getNetwork();
-  const networkId = NETWORKD_ID_NUMBER[network.name || network.id];
-  const rows = listEnabledPaymentAddresses(
-    Loader.Cardano,
-    account,
-    networkId
-  );
-  const hashes = [];
-  const seen = new Set();
-  const push = (h) => {
-    if (!h || seen.has(h)) return;
-    seen.add(h);
-    hashes.push(h);
-  };
-  push(account.paymentKeyHash);
-  for (const row of rows) push(row.paymentKeyHash);
-  return hashes;
-};
-
-/**
- * Persist which external address indices are active for the current account.
- * Index 0 is always kept. Triggers balance cache invalidation.
- */
-export const setAccountExternalIndices = async (indices) => {
-  const currentIndex = await getCurrentAccountIndex();
-  return setAccountExternalIndicesAt(currentIndex, indices);
-};
-
-/**
- * Persist external indices for a specific account storage slot.
- */
-export const setAccountExternalIndicesAt = async (accountIndex, indices) => {
-  const accounts = await getStorage(STORAGE.accounts);
-  if (!accounts?.[accountIndex]) {
-    throw new Error('Account not found');
-  }
-  const next = normalizeExternalIndices(indices);
-  accounts[accountIndex].externalIndices = next;
-  await setStorage({ [STORAGE.accounts]: { ...accounts } });
-  invalidateReadCache();
-  return next;
-};
-
-/**
- * Persist internal (change) indices for a specific account storage slot.
- */
-export const setAccountInternalIndicesAt = async (accountIndex, indices) => {
-  const accounts = await getStorage(STORAGE.accounts);
-  if (!accounts?.[accountIndex]) {
-    throw new Error('Account not found');
-  }
-  const next = normalizeInternalIndices(indices);
-  accounts[accountIndex].internalIndices = next;
-  await setStorage({ [STORAGE.accounts]: { ...accounts } });
-  invalidateReadCache();
-  return next;
-};
-
-/**
- * Networks to scan for used payment addresses under a stake key.
- * Preview/preprod share address bech32 format but are separate chains.
- */
-const EXTERNAL_ADDRESS_SCAN_NETWORKS = [
-  NETWORK_ID.mainnet,
-  NETWORK_ID.preview,
-  NETWORK_ID.preprod,
-];
-
-/** Consecutive empty payment indices before stopping address_txs fallback. */
-const EXTERNAL_ADDRESS_SCAN_GAP = 5;
-
-/**
- * Query Koios/Blockfrost for payment addresses on a stake key, then match
- * them to CIP-1852 external (role=0) and internal/change (role=1) indices.
- *
- * @param {object} account
- * @param {{ networkKeys?: string[] }} [options]
- * @returns {{ externalIndices: number[], internalIndices: number[] }}
- */
-export const discoverUsedPaymentIndices = async (account, options = {}) => {
-  await Loader.load();
-  const empty = {
-    externalIndices: getExternalIndices(account),
-    internalIndices: getInternalIndices(account),
-  };
-  if (!account?.publicKey) {
-    return {
-      externalIndices: [0],
-      internalIndices: [],
-    };
-  }
-
-  // Unit tests create wallets without a live chain; skip network discovery
-  // unless explicitly opted in (avoids slow/flaky CI).
-  if (
-    typeof process !== 'undefined' &&
-    process.env.JEST_WORKER_ID != null &&
-    process.env.LUCEM_DISCOVER_ADDRESSES !== '1'
-  ) {
-    return empty;
-  }
-
-  const networkKeys =
-    Array.isArray(options.networkKeys) && options.networkKeys.length > 0
-      ? options.networkKeys
-      : EXTERNAL_ADDRESS_SCAN_NETWORKS;
-
-  const externalFound = new Set([0]);
-  const internalFound = new Set();
-
-  for (const networkKey of networkKeys) {
-    const rewardAddr = account[networkKey]?.rewardAddr;
-    if (!rewardAddr) continue;
-    const networkIdNumber = NETWORKD_ID_NUMBER[networkKey];
-
-    try {
-      const req = KOIOS_REQUESTS.getAccountAddresses(rewardAddr, true);
-      const payload = await koiosRequest(
-        req.endpoint,
-        undefined,
-        req.body,
-        undefined,
-        networkKey
-      );
-      const addresses = flattenAccountAddressesPayload(payload);
-      for (const i of matchExternalIndicesFromAddresses(
-        Loader.Cardano,
-        account.publicKey,
-        networkIdNumber,
-        addresses
-      )) {
-        externalFound.add(i);
-      }
-      for (const i of matchInternalIndicesFromAddresses(
-        Loader.Cardano,
-        account.publicKey,
-        networkIdNumber,
-        addresses
-      )) {
-        internalFound.add(i);
-      }
-      continue;
-    } catch (error) {
-      console.warn(
-        `account_addresses scan failed (${networkKey}):`,
-        error.message || error
-      );
-    }
-
-    // Fallback when /account_addresses is unavailable: gap-limited /address_txs
-    // on both external and internal chains.
-    for (const role of [ADDRESS_ROLE.external, ADDRESS_ROLE.internal]) {
-      try {
-        let gap = 0;
-        const start = role === ADDRESS_ROLE.external ? 1 : 0;
-        const max =
-          role === ADDRESS_ROLE.external
-            ? MAX_EXTERNAL_ADDRESS_INDEX
-            : MAX_INTERNAL_ADDRESS_INDEX;
-        for (let i = start; i <= max; i++) {
-          const { paymentAddr } = derivePaymentFromAccountPublicKey(
-            Loader.Cardano,
-            account.publicKey,
-            networkIdNumber,
-            role,
-            i
-          );
-          const req = KOIOS_REQUESTS.getAddressTxs(paymentAddr);
-          const txs = await koiosRequest(
-            req.endpoint,
-            undefined,
-            req.body,
-            undefined,
-            networkKey
-          );
-          if (addressTxsIndicatesHistory(txs)) {
-            if (role === ADDRESS_ROLE.external) externalFound.add(i);
-            else internalFound.add(i);
-            gap = 0;
-          } else {
-            gap += 1;
-            if (gap >= EXTERNAL_ADDRESS_SCAN_GAP) break;
-          }
-        }
-      } catch (error) {
-        console.warn(
-          `address_txs role=${role} scan failed (${networkKey}):`,
-          error.message || error
-        );
-      }
-    }
-  }
-
-  return {
-    externalIndices: normalizeExternalIndices([...externalFound]),
-    internalIndices: normalizeInternalIndices([...internalFound]),
-  };
-};
-
-/** @deprecated Prefer discoverUsedPaymentIndices (includes change addresses). */
-export const discoverUsedExternalIndices = async (account, options = {}) => {
-  const { externalIndices } = await discoverUsedPaymentIndices(account, options);
-  return externalIndices;
-};
-
-/**
- * After import/create/refresh: discover used external + internal addresses
- * and activate them (merged with any already enabled indices).
- *
- * @param {string|number} accountIndex
- * @param {{ networkKeys?: string[] }} [options]
- * @returns {{ externalIndices: number[], internalIndices: number[] }}
- */
-export const activateDiscoveredExternalAddresses = async (
-  accountIndex,
-  options = {}
-) => {
-  const accounts = await getStorage(STORAGE.accounts);
-  const account = accounts?.[accountIndex];
-  if (!account) {
-    return { externalIndices: [0], internalIndices: [] };
-  }
-
-  try {
-    // Snapshot user-activated externals before discovery expands the enabled
-    // set, so empty discovered addresses stay off the Accounts address list.
-    if (!Array.isArray(accounts[accountIndex].userExternalIndices)) {
-      accounts[accountIndex].userExternalIndices = getExternalIndices(account);
-    }
-    const discovered = await discoverUsedPaymentIndices(account, options);
-    const mergedExternal = normalizeExternalIndices([
-      ...getExternalIndices(account),
-      ...discovered.externalIndices,
-    ]);
-    const mergedInternal = normalizeInternalIndices([
-      ...getInternalIndices(account),
-      ...discovered.internalIndices,
-    ]);
-    const prevExternal = getExternalIndices(account);
-    const prevInternal = getInternalIndices(account);
-    const externalSame =
-      mergedExternal.length === prevExternal.length &&
-      mergedExternal.every((n, i) => n === prevExternal[i]);
-    const internalSame =
-      mergedInternal.length === prevInternal.length &&
-      mergedInternal.every((n, i) => n === prevInternal[i]);
-    if (externalSame && internalSame) {
-      await setStorage({ [STORAGE.accounts]: { ...accounts } });
-      return {
-        externalIndices: mergedExternal,
-        internalIndices: mergedInternal,
-      };
-    }
-    accounts[accountIndex].externalIndices = mergedExternal;
-    accounts[accountIndex].internalIndices = mergedInternal;
-    await setStorage({ [STORAGE.accounts]: { ...accounts } });
-    invalidateReadCache();
-    return {
-      externalIndices: mergedExternal,
-      internalIndices: mergedInternal,
-    };
-  } catch (error) {
-    console.warn(
-      'External address activation failed:',
-      error.message || error
-    );
-    return {
-      externalIndices: getExternalIndices(account),
-      internalIndices: getInternalIndices(account),
-    };
-  }
-};
-
-/**
- * Enable a single external index (0..MAX) on the current account.
- * Marks the index as user-activated for the Accounts address list.
- */
-export const enableExternalAddressIndex = async (addressIndex) => {
-  const currentIndex = await getCurrentAccountIndex();
-  const accounts = await getStorage(STORAGE.accounts);
-  const account = accounts?.[currentIndex];
-  if (!account) throw new Error('Account not found');
-  const i = parseInt(addressIndex, 10);
-  if (!Number.isFinite(i) || i < 0 || i > MAX_EXTERNAL_ADDRESS_INDEX) {
-    throw new Error(
-      `Address index must be between 0 and ${MAX_EXTERNAL_ADDRESS_INDEX}`
-    );
-  }
-  const nextExternal = normalizeExternalIndices([
-    ...getExternalIndices(account),
-    i,
-  ]);
-  const nextUser = normalizeExternalIndices([
-    ...getUserExternalIndices(account),
-    i,
-  ]);
-  accounts[currentIndex].externalIndices = nextExternal;
-  accounts[currentIndex].userExternalIndices = nextUser;
-  await setStorage({ [STORAGE.accounts]: { ...accounts } });
-  invalidateReadCache();
-  return nextExternal;
-};
-
-/**
- * Disable an external index on the current account (index 0 cannot be disabled).
- * Also clears the user-activated flag for that index.
- */
-export const disableExternalAddressIndex = async (addressIndex) => {
-  const i = parseInt(addressIndex, 10);
-  if (i === 0) {
-    throw new Error('The primary address (index 0) cannot be disabled');
-  }
-  const currentIndex = await getCurrentAccountIndex();
-  const accounts = await getStorage(STORAGE.accounts);
-  const account = accounts?.[currentIndex];
-  if (!account) throw new Error('Account not found');
-  const nextExternal = normalizeExternalIndices(
-    getExternalIndices(account).filter((n) => n !== i)
-  );
-  const nextUser = normalizeExternalIndices(
-    getUserExternalIndices(account).filter((n) => n !== i)
-  );
-  accounts[currentIndex].externalIndices = nextExternal;
-  accounts[currentIndex].userExternalIndices = nextUser;
-  await setStorage({ [STORAGE.accounts]: { ...accounts } });
-  invalidateReadCache();
-  return nextExternal;
-};
-
-export const getRewardAddress = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  // Return the full Bech32 stake address instead of converting to key hash
-  return currentAccount.rewardAddr;
-};
-
-export const getPubDRepKey = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  if (!currentAccount?.publicKey) {
-    throw APIError.InternalError;
-  }
-  return deriveAccountDRepPublicKeyHex(currentAccount.publicKey);
-};
-
-/**
- * Derive this account's DRep identity (CIP-105 role-3 key).
- * Returns the raw key-hash hex (used as the voting credential / required signer)
- * plus the CIP-129 and legacy CIP-105 bech32 ids used to query registration.
- */
-export const getAccountDRepId = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  if (!currentAccount?.publicKey) {
-    throw APIError.InternalError;
-  }
-  const drepRawKey = Loader.Cardano.Bip32PublicKey.from_hex(currentAccount.publicKey)
-    .derive(3)
-    .derive(0)
-    .to_raw_key();
-  const drepKeyHash = drepRawKey.hash();
-  const drepKeyHashHex = Buffer.from(drepKeyHash.to_bytes()).toString('hex');
-  const drep = Loader.Cardano.DRep.new_key_hash(drepKeyHash);
-  let drepIdCip129 = '';
-  let drepIdLegacy = '';
-  try {
-    drepIdCip129 = drep.to_bech32(true);
-  } catch (e) {
-    drepIdCip129 = '';
-  }
-  try {
-    drepIdLegacy = drep.to_bech32(false);
-  } catch (e) {
-    drepIdLegacy = '';
-  }
-  return { drepKeyHashHex, drepIdCip129, drepIdLegacy };
-};
-
-export const getRegisteredPubStakeKeys = async () => {
-  await Loader.load();
-  const currentAccount = await getCurrentAccount();
-  if (!currentAccount?.publicKey || !currentAccount?.rewardAddr) {
-    throw APIError.InternalError;
-  }
-
-  const stakePubKeyHex = deriveAccountStakePublicKeyHex(currentAccount.publicKey);
-  const request = KOIOS_REQUESTS.getAccountInfo(currentAccount.rewardAddr);
-  const result = await koiosRequest(request.endpoint, {}, request.body);
-  const row = Array.isArray(result) && result.length > 0 ? result[0] : null;
-  const status = String(row?.status || '').toLowerCase();
-  const isRegistered =
-    row != null &&
-    (row.registered === true ||
-      (!status.includes('unreg') &&
-        !status.includes('not registered') &&
-        !status.includes('deregistered')));
-
-  return isRegistered ? [stakePubKeyHex] : [];
-};
-
-export const getUnregisteredPubStakeKeys = async () => {
-  const registeredKeys = await getRegisteredPubStakeKeys();
-  if (registeredKeys.length > 0) {
-    return [];
-  }
-  const currentAccount = await getCurrentAccount();
-  if (!currentAccount?.publicKey) {
-    throw APIError.InternalError;
-  }
-  return [deriveAccountStakePublicKeyHex(currentAccount.publicKey)];
-};
-
-export const getCurrentAccountIndex = () => getStorage(STORAGE.currentAccount);
-
-export const getNetwork = () => getStorage(STORAGE.network);
-
-export const setNetwork = async (network) => {
-  const currentNetwork = await getNetwork();
-  let id;
-  let node;
-  if (network.id === NETWORK_ID.mainnet) {
-    id = NETWORK_ID.mainnet;
-    node = NODE.mainnet;
-  } else if (network.id === NETWORK_ID.testnet) {
-    id = NETWORK_ID.testnet;
-    node = NODE.testnet;
-  } else if (network.id === NETWORK_ID.preview) {
-    id = NETWORK_ID.preview;
-    node = NODE.preview;
-  } else {
-    id = NETWORK_ID.preprod;
-    node = NODE.preprod;
-  }
-  if (network.node) node = network.node;
-  if (currentNetwork && currentNetwork.id !== id)
-    emitNetworkChange(networkNameToId(id));
-  await setStorage({
-    [STORAGE.network]: {
-      id,
-      node,
-      mainnetSubmit: network.mainnetSubmit,
-      testnetSubmit: network.testnetSubmit,
-    },
-  });
-  return true;
-};
-
-const accountToNetworkSpecific = (account, network) => {
-  const assets = account[network.id].assets;
-  const lovelace = account[network.id].lovelace;
-  const history = account[network.id].history;
-  const minAda = account[network.id].minAda;
-  const collateral = account[network.id].collateral;
-  const recentSendToAddresses = account[network.id].recentSendToAddresses;
-  const paymentAddr = account[network.id].paymentAddr;
-  const rewardAddr = account[network.id].rewardAddr;
-
-  return {
-    ...account,
-    paymentAddr,
-    rewardAddr,
-    assets,
-    lovelace,
-    minAda,
-    collateral,
-    history,
-    recentSendToAddresses,
-  };
-};
-
-/** Returns account with network specific settings (e.g. address, reward address, etc.) */
-export const getCurrentAccount = async () => {
-  const currentAccountIndex = await getCurrentAccountIndex();
-  const accounts = await getStorage(STORAGE.accounts);
-  const network = await getNetwork();
-  return accountToNetworkSpecific(accounts[currentAccountIndex], network);
-};
-
-/** True when encrypted storage has at least one account (wallet bootstrap / routing). */
-export const hasStoredAccounts = async () => {
-  const accounts = await getStorage(STORAGE.accounts);
-  return (
-    accounts != null &&
-    typeof accounts === 'object' &&
-    Object.keys(accounts).length > 0
-  );
-};
-
-/** Returns accounts with network specific settings (e.g. address, reward address, etc.) */
-export const getAccounts = async () => {
-  const accounts = await getStorage(STORAGE.accounts);
-  if (!accounts || typeof accounts !== 'object') {
-    return {};
-  }
-  const network = await getNetwork();
-  for (const index in accounts) {
-    accounts[index] = await accountToNetworkSpecific(accounts[index], network);
-  }
-  return accounts;
-};
 
 export const setAccountName = async (name) => {
   const currentAccountIndex = await getCurrentAccountIndex();
@@ -1848,9 +276,6 @@ export const openKeystoneSignTxTab = async ({ txHex, keyHashes, partialSign }) =
 export const getCurrentWebpage = () =>
   platform.navigation.getCurrentWebpage();
 
-const harden = (num) => {
-  return 0x80000000 + num;
-};
 
 export const bytesAddressToBinary = (bytes) =>
   bytes.reduce((str, byte) => str + byte.toString(2).padStart(8, '0'), '');
@@ -1899,634 +324,6 @@ export const isValidAddress = async (address) => {
   return false;
 };
 
-const isValidAddressBytes = async (address) => {
-  await Loader.load();
-  const network = await getNetwork();
-  try {
-    const addr = Loader.Cardano.Address.from_bytes(address);
-    if (
-      (addr.network_id() === 1 && network.id === NETWORK_ID.mainnet) ||
-      (addr.network_id() === 0 &&
-        (network.id === NETWORK_ID.testnet ||
-          network.id === NETWORK_ID.preview ||
-          network.id === NETWORK_ID.preprod))
-    )
-      return true;
-    return false;
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.ByronAddress.from_bytes(address);
-    if (
-      (addr.network_id() === 1 && network.id === NETWORK_ID.mainnet) ||
-      (addr.network_id() === 0 &&
-        (network.id === NETWORK_ID.testnet ||
-          network.id === NETWORK_ID.preview ||
-          network.id === NETWORK_ID.preprod))
-    )
-      return true;
-    return false;
-  } catch (e) {}
-  return false;
-};
-
-export const isValidEthAddress = function (address) {
-  return isAddress(address);
-};
-
-const DREP_ID_HEX_RE = /^[0-9a-f]{56}$/i;
-
-export const extractKeyHash = async (address) => {
-  await Loader.load();
-  if (DREP_ID_HEX_RE.test(address)) {
-    return `drep_vkh${address.toLowerCase()}`;
-  }
-  if (!(await isValidAddressBytes(Buffer.from(address, 'hex'))))
-    throw DataSignError.InvalidFormat;
-  try {
-    const addr = Loader.Cardano.BaseAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    return addr.payment_cred().to_keyhash().to_bech32('addr_vkh');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.EnterpriseAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    return addr.payment_cred().to_keyhash().to_bech32('addr_vkh');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.PointerAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    return addr.payment_cred().to_keyhash().to_bech32('addr_vkh');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.RewardAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    return addr.payment_cred().to_keyhash().to_bech32('stake_vkh');
-  } catch (e) {}
-  throw DataSignError.AddressNotPK;
-};
-
-const deriveAccountDRepPrivateKey = (accountKey) => accountKey.derive(3).derive(0).to_raw_key();
-
-const deriveAccountStakePublicKeyHex = (accountPublicKeyHex) => {
-  const stakeKey = Loader.Cardano.Bip32PublicKey.from_hex(accountPublicKeyHex)
-    .derive(2)
-    .derive(0)
-    .to_raw_key();
-  return Buffer.from(stakeKey.as_bytes()).toString('hex');
-};
-
-const deriveAccountDRepPublicKeyHex = (accountPublicKeyHex) => {
-  const drepKey = Loader.Cardano.Bip32PublicKey.from_hex(accountPublicKeyHex)
-    .derive(3)
-    .derive(0)
-    .to_raw_key();
-  return Buffer.from(drepKey.as_bytes()).toString('hex');
-};
-
-export const extractKeyOrScriptHash = async (address) => {
-  console.log('extractKeyOrScriptHash', address);
-  await Loader.load();
-  if (!(await isValidAddressBytes(Buffer.from(address, 'hex'))))
-    throw DataSignError.InvalidFormat;
-  try {
-    const addr = Loader.Cardano.BaseAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-
-    const credential = addr.payment_cred();
-    if (credential.kind() === 0)
-      return credential.to_keyhash().to_bech32('addr_vkh');
-    if (credential.kind() === 1)
-      return credential.to_scripthash().to_bech32('script');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.EnterpriseAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    const credential = addr.payment_cred();
-    if (credential.kind() === 0)
-      return credential.to_keyhash().to_bech32('addr_vkh');
-    if (credential.kind() === 1)
-      return credential.to_scripthash().to_bech32('script');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.PointerAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    const credential = addr.payment_cred();
-    if (credential.kind() === 0)
-      return credential.to_keyhash().to_bech32('addr_vkh');
-    if (credential.kind() === 1)
-      return credential.to_scripthash().to_bech32('script');
-  } catch (e) {}
-  try {
-    const addr = Loader.Cardano.RewardAddress.from_address(
-      Loader.Cardano.Address.from_bytes(Buffer.from(address, 'hex'))
-    );
-    const credential = addr.payment_cred();
-    if (credential.kind() === 0)
-      return credential.to_keyhash().to_bech32('stake_vkh');
-    if (credential.kind() === 1)
-      return credential.to_scripthash().to_bech32('script');
-  } catch (e) {}
-  throw new Error('No address type matched.');
-};
-
-export const verifySigStructure = async (sigStructure) => {
-  await Loader.load();
-  try {
-    Loader.Message.SigStructure.from_bytes(Buffer.from(sigStructure, 'hex'));
-  } catch (e) {
-    throw DataSignError.InvalidFormat;
-  }
-};
-
-export const verifyPayload = (payload) => {
-  if (Buffer.from(payload, 'hex').length <= 0)
-    throw DataSignError.InvalidFormat;
-};
-
-export const verifyTx = async (tx) => {
-  await Loader.load();
-  const network = await getNetwork();
-  try {
-    const parseTx = Loader.Cardano.Transaction.from_bytes(Buffer.from(tx, 'hex'));
-    let networkId = parseTx.body().network_id()
-      ? parseTx.body().network_id().network()
-      : null;
-    if (!networkId && networkId != 0) {
-      networkId = parseTx.body().outputs().get(0).address().network_id();
-    }
-    if (networkId != networkNameToId(network.id)) throw Error('Wrong network');
-  } catch (e) {
-    throw APIError.InvalidRequest;
-  }
-};
-
-/**
- * @param {string} address - cbor
- * @param {string} payload - hex encoded utf8 string
- * @param {string} password
- * @param {number} accountIndex
- * @returns
- */
-
-//deprecated soon
-export const signData = async (address, payload, password, accountIndex) => {
-  await Loader.load();
-  const keyHash = await extractKeyHash(address);
-  const prefix = keyHash.startsWith('addr_vkh')
-    ? 'addr_vkh'
-    : keyHash.startsWith('drep_vkh')
-      ? 'drep_vkh'
-      : 'stake_vkh';
-  const sdAccounts = await getStorage(STORAGE.accounts);
-  const sdAccount = sdAccounts?.[accountIndex];
-  let { accountKey, paymentKey, stakeKey } = await requestAccountKey(
-    password,
-    sdAccount?.derivationIndex ?? accountIndex,
-    sdAccount?.walletId ?? null
-  );
-  let drepKey = deriveAccountDRepPrivateKey(accountKey);
-  const signingKey =
-    prefix === 'addr_vkh' ? paymentKey : prefix === 'drep_vkh' ? drepKey : stakeKey;
-
-  const publicKey = signingKey.to_public();
-  if (keyHash !== publicKey.hash().to_bech32(prefix))
-    throw DataSignError.ProofGeneration;
-
-  const protectedHeaders = Loader.Message.HeaderMap.new();
-  protectedHeaders.set_algorithm_id(
-    Loader.Message.Label.from_algorithm_id(Loader.Message.AlgorithmId.EdDSA)
-  );
-  protectedHeaders.set_key_id(publicKey.as_bytes());
-  protectedHeaders.set_header(
-    Loader.Message.Label.new_text('address'),
-    Loader.Message.CBORValue.new_bytes(Buffer.from(address, 'hex'))
-  );
-  const protectedSerialized =
-    Loader.Message.ProtectedHeaderMap.new(protectedHeaders);
-  const unprotectedHeaders = Loader.Message.HeaderMap.new();
-  const headers = Loader.Message.Headers.new(
-    protectedSerialized,
-    unprotectedHeaders
-  );
-  const builder = Loader.Message.COSESign1Builder.new(
-    headers,
-    Buffer.from(payload, 'hex'),
-    false
-  );
-  const toSign = builder.make_data_to_sign().to_bytes();
-
-  const signedSigStruc = signingKey.sign(toSign).to_bytes();
-  const coseSign1 = builder.build(signedSigStruc);
-
-  accountKey.free();
-  accountKey = null;
-  drepKey.free();
-  drepKey = null;
-  stakeKey.free();
-  stakeKey = null;
-  paymentKey.free();
-  paymentKey = null;
-
-  return Buffer.from(coseSign1.to_bytes(), 'hex').toString('hex');
-};
-
-export const signDataCIP30 = async (
-  address,
-  payload,
-  password,
-  accountIndex
-) => {
-  await Loader.load();
-  const keyHash = await extractKeyHash(address);
-  const prefix = keyHash.startsWith('addr_vkh')
-    ? 'addr_vkh'
-    : keyHash.startsWith('drep_vkh')
-      ? 'drep_vkh'
-      : 'stake_vkh';
-  const cip30Accounts = await getStorage(STORAGE.accounts);
-  const cip30Account = cip30Accounts?.[accountIndex];
-  let { accountKey, paymentKey, stakeKey } = await requestAccountKey(
-    password,
-    cip30Account?.derivationIndex ?? accountIndex,
-    cip30Account?.walletId ?? null
-  );
-  let drepKey = deriveAccountDRepPrivateKey(accountKey);
-  const signingKey =
-    prefix === 'addr_vkh' ? paymentKey : prefix === 'drep_vkh' ? drepKey : stakeKey;
-
-  const publicKey = signingKey.to_public();
-  if (keyHash !== publicKey.hash().to_bech32(prefix))
-    throw DataSignError.ProofGeneration;
-  const protectedHeaders = Loader.Message.HeaderMap.new();
-  protectedHeaders.set_algorithm_id(
-    Loader.Message.Label.from_algorithm_id(Loader.Message.AlgorithmId.EdDSA)
-  );
-  // protectedHeaders.set_key_id(publicKey.to_raw_bytes()); // Removed to adhere to CIP-30
-  protectedHeaders.set_header(
-    Loader.Message.Label.new_text('address'),
-    Loader.Message.CBORValue.new_bytes(
-      Buffer.from(
-        prefix === 'drep_vkh' ? publicKey.hash().to_hex() : address,
-        'hex'
-      )
-    )
-  );
-  const protectedSerialized =
-    Loader.Message.ProtectedHeaderMap.new(protectedHeaders);
-  const unprotectedHeaders = Loader.Message.HeaderMap.new();
-  const headers = Loader.Message.Headers.new(
-    protectedSerialized,
-    unprotectedHeaders
-  );
-  const builder = Loader.Message.COSESign1Builder.new(
-    headers,
-    Buffer.from(payload, 'hex'),
-    false
-  );
-  const toSign = builder.make_data_to_sign().to_bytes();
-
-  const signedSigStruc = signingKey.sign(toSign).to_bytes();
-  const coseSign1 = builder.build(signedSigStruc);
-
-  accountKey.free();
-  accountKey = null;
-  drepKey.free();
-  drepKey = null;
-  stakeKey.free();
-  stakeKey = null;
-  paymentKey.free();
-  paymentKey = null;
-
-  const key = Loader.Message.COSEKey.new(
-    Loader.Message.Label.from_key_type(Loader.Message.KeyType.OKP)
-  );
-  key.set_algorithm_id(
-    Loader.Message.Label.from_algorithm_id(Loader.Message.AlgorithmId.EdDSA)
-  );
-  key.set_header(
-    Loader.Message.Label.new_int(
-      Loader.Message.Int.new_negative(Loader.Message.BigNum.from_str('1'))
-    ),
-    Loader.Message.CBORValue.new_int(
-      Loader.Message.Int.new_i32(6) //Loader.Message.CurveType.Ed25519
-    )
-  ); // crv (-1) set to Ed25519 (6)
-  key.set_header(
-    Loader.Message.Label.new_int(
-      Loader.Message.Int.new_negative(Loader.Message.BigNum.from_str('2'))
-    ),
-    Loader.Message.CBORValue.new_bytes(publicKey.as_bytes())
-  ); // x (-2) set to public key
-
-  return {
-    signature: Buffer.from(coseSign1.to_bytes()).toString('hex'),
-    key: Buffer.from(key.to_bytes()).toString('hex'),
-  };
-};
-
-/**
- *
- * @param {string} tx - cbor hex string
- * @param {Array<string>} keyHashes
- * @param {string} password
- * @returns {Promise<string>} witness set as hex string
- */
-export const signTx = async (
-  tx,
-  keyHashes,
-  password,
-  accountIndex,
-  partialSign = false
-) => {
-  await Loader.load();
-  // `accountIndex` is the storage slot. Resolve the seed + CIP-1852 index it maps
-  // to so multi-seed accounts sign with the correct root key (legacy accounts
-  // fall back to slot == derivation index, walletId "0").
-  const accounts = await getStorage(STORAGE.accounts);
-  const account = accounts?.[accountIndex];
-  const derivationIndex = account?.derivationIndex ?? accountIndex;
-  const walletId = account?.walletId ?? null;
-  let { accountKey, paymentKey, stakeKey } = await requestAccountKey(
-    password,
-    derivationIndex,
-    walletId
-  );
-  let drepKey = deriveAccountDRepPrivateKey(accountKey);
-  const paymentKeyHash = paymentKey.to_public().hash().to_hex();
-  const stakeKeyHash = stakeKey.to_public().hash().to_hex();
-  const drepKeyHash = drepKey.to_public().hash().to_hex();
-
-  const keyMap = new Map([
-    [paymentKeyHash, paymentKey],
-    [stakeKeyHash, stakeKey],
-    [drepKeyHash, drepKey],
-  ]);
-
-  // Advanced multi-address: include payment keys for every enabled external and
-  // internal (change) index so inputs on those addresses can be witnessed.
-  const extraPaymentKeys = [];
-  for (const addressIndex of getExternalIndices(account).filter((i) => i !== 0)) {
-    const extraKey = accountKey.derive(0).derive(addressIndex).to_raw_key();
-    extraPaymentKeys.push(extraKey);
-    keyMap.set(extraKey.to_public().hash().to_hex(), extraKey);
-  }
-  for (const addressIndex of getInternalIndices(account)) {
-    const extraKey = accountKey.derive(1).derive(addressIndex).to_raw_key();
-    extraPaymentKeys.push(extraKey);
-    keyMap.set(extraKey.to_public().hash().to_hex(), extraKey);
-  }
-
-  let txWitnessSet;
-  try {
-    txWitnessSet = buildVkeyWitnessSet(
-      Loader.Cardano,
-      tx,
-      keyMap,
-      keyHashes,
-      partialSign
-    );
-  } catch {
-    throw TxSignError.ProofGeneration;
-  } finally {
-    accountKey.free();
-    drepKey.free();
-    stakeKey.free();
-    paymentKey.free();
-    extraPaymentKeys.forEach((k) => {
-      try {
-        k.free();
-      } catch (_) {
-        /* ignore */
-      }
-    });
-  }
-
-  return txWitnessSet;
-};
-
-export const signTxHW = async (
-  tx,
-  keyHashes,
-  account,
-  hw,
-  partialSign = false
-) => {
-  await Loader.load();
-  const rawTx = Loader.Cardano.Transaction.from_bytes(Buffer.from(tx, 'hex'));
-  const address = Loader.Cardano.Address.from_bech32(account.paymentAddr);
-  const network = address.network_id();
-  const keys = {
-    payment: { hash: null, path: null },
-    stake: { hash: null, path: null },
-  };
-  if (hw.device === HW.ledger) {
-    const appAda = hw.appAda;
-    const networkId = network;
-    const paymentIndexByHash = {};
-    if (account?.publicKey) {
-      for (const row of listEnabledPaymentAddresses(
-        Loader.Cardano,
-        account,
-        networkId
-      )) {
-        paymentIndexByHash[row.paymentKeyHash] = {
-          index: row.index,
-          role: row.role ?? ADDRESS_ROLE.external,
-        };
-      }
-    } else {
-      paymentIndexByHash[account.paymentKeyHash] = {
-        index: 0,
-        role: ADDRESS_ROLE.external,
-      };
-    }
-    keyHashes.forEach((keyHash) => {
-      if (paymentIndexByHash[keyHash] != null) {
-        const { index: addrIdx, role } = paymentIndexByHash[keyHash];
-        keys.payment = {
-          hash: keyHash,
-          path: [
-            HARDENED + 1852,
-            HARDENED + 1815,
-            HARDENED + hw.account,
-            role,
-            addrIdx,
-          ],
-        };
-      } else if (keyHash === account.stakeKeyHash)
-        keys.stake = {
-          hash: keyHash,
-          path: [HARDENED + 1852, HARDENED + 1815, HARDENED + hw.account, 2, 0],
-        };
-      else if (!partialSign) throw TxSignError.ProofGeneration;
-      else return;
-    });
-    const ledgerTx = await txToLedger(
-      rawTx,
-      network,
-      keys,
-      Buffer.from(address.to_bytes()).toString('hex'),
-      hw.account
-    );
-    const result = await appAda.signTransaction({
-      ...ledgerTx,
-      options: {
-        tagCborSets: hasTaggedSets(tx)
-      }
-    });
-    // getting public keys
-    const witnessSet = Loader.Cardano.TransactionWitnessSet.new();
-    const vkeys = Loader.Cardano.Vkeywitnesses.new();
-    result.witnesses.forEach((witness) => {
-      const role = witness.path[3];
-      if (role === 0 || role === 1) {
-        const addrIdx = witness.path[4] != null ? witness.path[4] : 0;
-        const vkey = Loader.Cardano.Bip32PublicKey.from_hex(
-          account.publicKey
-        )
-          .derive(role)
-          .derive(addrIdx)
-          .to_raw_key();
-        const signature = Loader.Cardano.Ed25519Signature.from_hex(
-          witness.witnessSignatureHex
-        );
-        vkeys.add(Loader.Cardano.Vkeywitness.new(vkey, signature));
-      } else if (
-        role == 2 // stake key
-      ) {
-        const vkey = Loader.Cardano.Bip32PublicKey.from_hex(
-          account.publicKey
-        )
-          .derive(2)
-          .derive(0)
-          .to_raw_key();
-        const signature = Loader.Cardano.Ed25519Signature.from_hex(
-          witness.witnessSignatureHex
-        );
-        vkeys.add(Loader.Cardano.Vkeywitness.new(vkey, signature));
-      }
-    });
-    witnessSet.set_vkeys(vkeys);
-    return witnessSet;
-  }
-  if (hw.device === HW.keystone) {
-    throw new Error('Keystone signing runs in the Keystone signing tab.');
-  }
-  if (hw.device === HW.trezor) {
-    const paymentIndexByHash = {};
-    if (account?.publicKey) {
-      for (const row of listEnabledPaymentAddresses(
-        Loader.Cardano,
-        account,
-        network
-      )) {
-        paymentIndexByHash[row.paymentKeyHash] = {
-          index: row.index,
-          role: row.role ?? ADDRESS_ROLE.external,
-        };
-      }
-    } else {
-      paymentIndexByHash[account.paymentKeyHash] = {
-        index: 0,
-        role: ADDRESS_ROLE.external,
-      };
-    }
-    keyHashes.forEach((keyHash) => {
-      if (paymentIndexByHash[keyHash] != null) {
-        const { index: addrIdx, role } = paymentIndexByHash[keyHash];
-        keys.payment = {
-          hash: keyHash,
-          path: `m/1852'/1815'/${hw.account}'/${role}/${addrIdx}`,
-        };
-      } else if (keyHash === account.stakeKeyHash)
-        keys.stake = {
-          hash: keyHash,
-          path: `m/1852'/1815'/${hw.account}'/2/0`,
-        };
-      else if (!partialSign) throw TxSignError.ProofGeneration;
-      else return;
-    });
-    const trezorTx = await txToTrezor(
-      rawTx,
-      network,
-      keys,
-      Buffer.from(address.to_bytes()).toString('hex'),
-      hw.account
-    );
-    const result = await TrezorConnect.cardanoSignTransaction({
-      ...trezorTx,
-      tagCborSets: hasTaggedSets(tx),
-    });
-    if (!result.success) throw new Error('Trezor could not sign tx');
-    const witnessSet = Loader.Cardano.TransactionWitnessSet.new();
-    const vkeys = Loader.Cardano.Vkeywitnesses.new();
-    result.payload.witnesses.forEach((witness) => {
-      const vkey = Loader.Cardano.PublicKey.from_bytes(
-        Buffer.from(witness.pubKey, 'hex')
-      );
-      const signature = Loader.Cardano.Ed25519Signature.from_hex(
-        witness.signature
-      );
-      vkeys.add(Loader.Cardano.Vkeywitness.new(vkey, signature));
-    });
-    witnessSet.set_vkeys(vkeys);
-    return witnessSet;
-  }
-  throw new Error('Unsupported hardware wallet device');
-};
-
-/**
- *
- * @param {string} tx - cbor hex string
- * @returns
- */
-
-export const submitTx = async (tx) => {
-  const network = await getNetwork();
-  
-  // Convert CBOR to hex if needed
-  const txHex = typeof tx === 'string' ? tx : Buffer.from(tx).toString('hex');
-  
-  if (network[network.id + 'Submit']) {
-    const result = await fetch(network[network.id + 'Submit'], {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/cbor' },
-      body: nativeSafeBinaryBody(Buffer.from(txHex, 'hex'), 'application/cbor'),
-    });
-    if (result.ok) {
-      // Balance/UTxO/history caches are now stale — drop them so the next read
-      // reflects the just-submitted transaction.
-      invalidateReadCache();
-      return await result.json();
-    }
-    throw APIError.InvalidRequest;
-  }
-  
-  try {
-    const result = await koiosSubmitTransaction(txHex);
-    invalidateReadCache();
-    return result;
-  } catch (error) {
-    console.error('Koios transaction submission error:', error);
-    throw new Error(`Transaction submission failed: ${error.message}`);
-  }
-};
-
-const emitNetworkChange = async (networkId) => {
-  platform.events.broadcastToTabs({
-    data: networkId,
-    target: TARGET,
-    sender: SENDER.extension,
-    event: EVENT.networkChange,
-  });
-};
 
 const emitAccountChange = async (addresses) => {
   if (typeof window !== 'undefined') {
@@ -2573,93 +370,6 @@ export const switchAccount = async (accountIndex) => {
   await setStorage({ [STORAGE.currentAccount]: accountIndex });
   const address = await getAddress();
   emitAccountChange([address]);
-  return true;
-};
-
-/**
- * Resolve the encrypted root key for a given seed.
- *
- * `walletId` is a stable per-seed id. Legacy single-seed installs (and the very
- * first seed) live in `STORAGE.encryptedKey` and answer to walletId "0"; every
- * other seed lives in the `STORAGE.encryptedKeys` map written when a second seed
- * is added.
- */
-const resolveEncryptedRootKey = async (walletId) => {
-  const id = walletId == null || walletId === '' ? '0' : String(walletId);
-  const map = await getStorage(STORAGE.encryptedKeys);
-  if (map && typeof map === 'object' && map[id]) return map[id];
-  if (id === '0') {
-    const legacy = await getStorage(STORAGE.encryptedKey);
-    if (legacy) return legacy;
-  }
-  throw new Error(`No stored key for wallet ${id}`);
-};
-
-export const requestAccountKey = async (password, accountIndex, walletId = null) => {
-  await Loader.load();
-  const encryptedRootKey = await resolveEncryptedRootKey(walletId);
-  let accountKey;
-  let decryptedHex;
-  try {
-    decryptedHex = await decryptWithPassword(password, encryptedRootKey);
-  } catch (e) {
-    throw ERROR.wrongPassword;
-  }
-  try {
-    accountKey = Loader.Cardano.Bip32PrivateKey.from_bytes(
-      Buffer.from(decryptedHex, 'hex')
-    )
-      .derive(harden(1852)) // purpose
-      .derive(harden(1815)) // coin type
-      .derive(harden(parseInt(accountIndex)));
-  } catch (e) {
-    console.error('Key derivation failed after successful decryption:', e);
-    throw ERROR.wrongPassword;
-  }
-
-  return {
-    accountKey,
-    paymentKey: accountKey.derive(0).derive(0).to_raw_key(),
-    stakeKey: accountKey.derive(2).derive(0).to_raw_key(),
-  };
-};
-
-/**
- * Re-encrypt every seed in the vault (legacy `encryptedKey` + all `encryptedKeys`)
- * under a new password. Throws `ERROR.wrongPassword` if `currentPassword` is wrong.
- */
-export const changeWalletPassword = async (currentPassword, newPassword) => {
-  await Loader.load();
-
-  const reencrypt = async (encryptedHex) => {
-    let decryptedHex;
-    try {
-      decryptedHex = await decryptWithPassword(currentPassword, encryptedHex);
-    } catch (e) {
-      throw ERROR.wrongPassword;
-    }
-    let rootKey = Loader.Cardano.Bip32PrivateKey.from_bytes(
-      Buffer.from(decryptedHex, 'hex')
-    );
-    const out = await encryptWithPassword(newPassword, rootKey.as_bytes());
-    rootKey.free();
-    rootKey = null;
-    return out;
-  };
-
-  const legacy = await getStorage(STORAGE.encryptedKey);
-  const map = await getStorage(STORAGE.encryptedKeys);
-
-  const patch = {};
-  if (legacy) patch[STORAGE.encryptedKey] = await reencrypt(legacy);
-  if (map && typeof map === 'object' && Object.keys(map).length > 0) {
-    const nextMap = {};
-    for (const id of Object.keys(map)) {
-      nextMap[id] = await reencrypt(map[id]);
-    }
-    patch[STORAGE.encryptedKeys] = nextMap;
-  }
-  await setStorage(patch);
   return true;
 };
